review: separated file for sql requests
